admin/requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.9.3
bs4==0.0.2
certifi==2025.1.31
chardet==3.0.4
Django==2.2.25
django-cors-headers==3.4.0
djangorestframework==3.9.4
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
et-xmlfile==1.1.0
hdfs==2.7.3
idna==2.6
importlib-metadata==6.7.0
jieba==0.42.1
Jinja2==3.1.6
joblib==1.3.2
MarkupSafe==2.1.5
mysqlclient==1.4.6
numpy==1.21.6
openpyxl==3.1.3
pandas==1.3.5
pip==24.0
prettytable==3.7.0
py4j==0.10.9
pyecharts==2.0.8
PyHDFS==0.3.1
PyJWT==2.8.0
pyspark==3.0.3
python-dateutil==2.9.0.post0
pytz==2023.4
requests==2.18.4
scikit-learn==1.0.2
scipy==1.7.3
setuptools==47.1.0
simplejson==3.20.1
six==1.17.0
soupsieve==2.4.1
spark==0.2.1
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.7.1
urllib3==1.22
wcwidth==0.2.13
zipp==3.15.0